misc/refactor_form_stories
