+++ src/storybook/package.json
@@ -20,15 +20,15 @@
     },
     "devDependencies": {
         "@babel/core": "^7.6.0",
-        "@storybook/addon-a11y": "^5.2.1",
-        "@storybook/addon-actions": "^5.2.1",
-        "@storybook/addon-backgrounds": "^5.2.1",
-        "@storybook/addon-knobs": "^5.2.1",
-        "@storybook/addon-notes": "^5.2.1",
-        "@storybook/addon-storyshots": "^5.2.1",
-        "@storybook/addon-storysource": "^5.2.1",
-        "@storybook/addon-viewport": "^5.2.1",
-        "@storybook/vue": "^5.2.1",
+        "@storybook/addon-a11y": "^5.2.5",
+        "@storybook/addon-actions": "^5.2.5",
+        "@storybook/addon-backgrounds": "^5.2.5",
+        "@storybook/addon-knobs": "^5.2.5",
+        "@storybook/addon-notes": "^5.2.5",
+        "@storybook/addon-storyshots": "^5.2.5",
+        "@storybook/addon-storysource": "^5.2.5",
+        "@storybook/addon-viewport": "^5.2.5",
+        "@storybook/vue": "^5.2.5",
         "@types/storybook__addon-storyshots": "^5.1.1",
         "@types/webpack-env": "^1.13.6",
         "acorn": "^7.0.0",
